+++ crates/swc_ecma_lints/src/rules/prefer_regex_literals.rs
@@ -1,5 +1,10 @@
+use std::{
+    fmt::{self, Debug},
+    sync::Arc,
+};
+
 use serde::{Deserialize, Serialize};
-use swc_common::{collections::AHashSet, errors::HANDLER, Span, SyntaxContext};
+use swc_common::{collections::AHashSet, errors::HANDLER, SourceMap, Span, SyntaxContext};
 use swc_ecma_ast::*;
 use swc_ecma_utils::{collect_decls_with_ctxt, ident::IdentLike};
 use swc_ecma_visit::{noop_visit_type, Visit, VisitWith};
@@ -7,6 +12,7 @@ use swc_ecma_visit::{noop_visit_type, Visit, VisitWith};
 use crate::{
     config::{LintRuleReaction, RuleConfig},
     rule::{visitor_rule, Rule},
+    rules::utils::{extract_arg_val, ArgValue},
 };
 
 const UNEXPECTED_REG_EXP_MESSAGE: &str =
@@ -28,6 +34,7 @@ pub struct PreferRegexLiteralsConfig {
 
 pub fn prefer_regex_literals(
     program: &Program,
+    source_map: &Arc<SourceMap>,
     config: &RuleConfig<PreferRegexLiteralsConfig>,
     top_level_ctxt: SyntaxContext,
     es_version: EsVersion,
@@ -44,6 +51,7 @@ pub fn prefer_regex_literals(
     match rule_reaction {
         LintRuleReaction::Off => None,
         _ => Some(visitor_rule(PreferRegexLiterals::new(
+            source_map.clone(),
             rule_reaction,
             disallow_redundant_wrapping,
             top_level_declared_vars,
@@ -52,27 +60,41 @@ pub fn prefer_regex_literals(
         ))),
     }
 }
-#[derive(Debug)]
-enum ArgValueType {
-    Str,
-    RegExp,
-    Another,
-}
 
-#[derive(Debug, Default)]
+#[derive(Default)]
 struct PreferRegexLiterals {
     expected_reaction: LintRuleReaction,
+    source_map: Arc<SourceMap>,
     disallow_redundant_wrapping: bool,
     top_level_ctxt: SyntaxContext,
     top_level_declared_vars: AHashSet<Id>,
     allow_global_this: bool,
     call_span: Option<Span>,
-    first_arg: Option<ArgValueType>,
-    second_arg: Option<ArgValueType>,
+    first_arg: Option<ArgValue>,
+    second_arg: Option<ArgValue>,
+}
+
+impl Debug for PreferRegexLiterals {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("PreferRegexLiterals")
+            .field("expected_reaction", &self.expected_reaction)
+            .field(
+                "disallow_redundant_wrapping",
+                &self.disallow_redundant_wrapping,
+            )
+            .field("top_level_ctxt", &self.top_level_ctxt)
+            .field("top_level_declared_vars", &self.top_level_declared_vars)
+            .field("allow_global_this", &self.allow_global_this)
+            .field("call_span", &self.call_span)
+            .field("first_arg", &self.first_arg)
+            .field("second_arg", &self.second_arg)
+            .finish()
+    }
 }
 
 impl PreferRegexLiterals {
     fn new(
+        source_map: Arc<SourceMap>,
         expected_reaction: LintRuleReaction,
         disallow_redundant_wrapping: bool,
         top_level_declared_vars: AHashSet<Id>,
@@ -81,6 +103,7 @@ impl PreferRegexLiterals {
     ) -> Self {
         Self {
             expected_reaction,
+            source_map,
             disallow_redundant_wrapping,
             top_level_ctxt,
             top_level_declared_vars,
@@ -88,25 +111,6 @@ impl PreferRegexLiterals {
             call_span: None,
             first_arg: None,
             second_arg: None,
-        }
-    }
-
-    fn extract_arg_value_type(expr: &Expr) -> ArgValueType {
-        match expr {
-            Expr::Lit(Lit::Str(_)) => ArgValueType::Str,
-            Expr::Lit(Lit::Regex(_)) => ArgValueType::RegExp,
-            Expr::Tpl(Tpl { exprs, .. }) => match exprs.len() {
-                0 => ArgValueType::Str,
-                _ => ArgValueType::Another,
-            },
-            Expr::TaggedTpl(TaggedTpl {
-                tpl: Tpl { exprs, .. },
-                ..
-            }) => match exprs.len() {
-                0 => ArgValueType::Str,
-                _ => ArgValueType::Another,
-            },
-            _ => ArgValueType::Another,
         }
     }
 
@@ -120,11 +124,23 @@ impl PreferRegexLiterals {
         self.call_span = Some(call_span);
 
         if let Some(ExprOrSpread { expr, .. }) = args.get(0) {
-            self.first_arg = Some(Self::extract_arg_value_type(expr.as_ref()));
+            self.first_arg = Some(extract_arg_val(
+                &self.source_map,
+                &self.top_level_ctxt,
+                &self.top_level_declared_vars,
+                expr.as_ref(),
+                true,
+            ));
         }
 
         if let Some(ExprOrSpread { expr, .. }) = args.get(1) {
-            self.second_arg = Some(Self::extract_arg_value_type(expr.as_ref()));
+            self.second_arg = Some(extract_arg_val(
+                &self.source_map,
+                &self.top_level_ctxt,
+                &self.top_level_declared_vars,
+                expr.as_ref(),
+                true,
+            ));
         }
     }
 
@@ -143,18 +159,18 @@ impl PreferRegexLiterals {
     }
 
     fn check(&self, sym: &str) {
-        use ArgValueType::*;
+        use ArgValue::*;
 
         if sym == "RegExp" {
             match (self.first_arg.as_ref(), self.second_arg.as_ref()) {
-                (Some(Str), None) => self.emit_report(UNEXPECTED_REG_EXP_MESSAGE),
-                (Some(Str), Some(Str)) => self.emit_report(UNEXPECTED_REG_EXP_MESSAGE),
-                (Some(RegExp), None) => {
+                (Some(Str(_)), None) => self.emit_report(UNEXPECTED_REG_EXP_MESSAGE),
+                (Some(Str(_)), Some(Str(_))) => self.emit_report(UNEXPECTED_REG_EXP_MESSAGE),
+                (Some(RegExp { .. }), None) => {
                     if self.disallow_redundant_wrapping {
                         self.emit_report(UNEXPECTED_REDUNDANT_REG_EXP_MESSAGE);
                     }
                 }
-                (Some(RegExp), Some(Str)) => {
+                (Some(RegExp { .. }), Some(Str(_))) => {
                     if self.disallow_redundant_wrapping {
                         self.emit_report(UNEXPECTED_REDUNDANT_REG_EXP_WITH_FLAGS_MESSAGE);
                     }